master_agent.py
import os
import argparse
import json
from typing import Dict, List, Optional, Any
from pydantic import BaseModel, Field
from langgraph.graph import StateGraph, END
import gdown
import tempfile
from datetime import datetime

# Attempt to import Google API client libraries
try:
    from googleapiclient.discovery import build
    from google.oauth2.service_account import Credentials as ServiceAccountCredentials
    from google.auth.transport.requests import Request
    from google.oauth2.credentials import Credentials
    from google_auth_oauthlib.flow import InstalledAppFlow
    GOOGLE_API_AVAILABLE = True
except ImportError:
    GOOGLE_API_AVAILABLE = False
    print("WARNING: Google API client libraries not found. Folder processing will be limited.")

from utils import (
    initialize_master_json, 
    save_master_json, 
    get_audio_file_identifier,
    add_agent_output_to_master_json,
    extract_google_drive_file_id,
    scrape_google_drive_folder,
    MASTER_JSON_FILE
)

# Placeholder for refactored agents - these will be imported later
from transcription_diarization_agent import process_audio_for_transcription_diarization
from evaluation_agent_refactored import run_evaluation_analysis
from noise_analysis_agent import run_noise_analysis
from recommendation_agent_refactored import run_recommendation_generation
from summary_agent_refactored import run_summary_generation

# --- Master Agent State ---
class MasterAgentState(BaseModel):
    master_json_data: Dict[str, Any] = Field(default_factory=initialize_master_json)
    initial_google_drive_link: Optional[str] = None # The original link provided by the user
    current_processing_link: Optional[str] = None # Link of the file currently being processed
    current_file_identifier: Optional[str] = None
    current_local_audio_path: Optional[str] = None
    user_prompt: Optional[str] = None
    is_folder: bool = False # Flag to indicate if the link is a folder
    files_to_process_queue: List[Dict[str, str]] = Field(default_factory=list) # Queue of {'link': str, 'identifier': str}
    processed_files_count: int = 0
    total_files_to_process: int = 0
    overall_status: str = "pending"
    error_message: Optional[str] = None
# ... (rest of file omitted for brevity, but will be uploaded in full) ...
