requirements.txt
beautifulsoup4
dotenv
gdown
google-generativeai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langgraph
librosa
numpy
openai
pandas
pyannote.audio
pydantic
requests
torch
